PMM-8267 change status styling logic
